requirements.txt
# Core framework
streamlit>=1.28.0
fastapi>=0.104.0

# Document processing
pdfplumber>=0.7.6
python-docx>=0.8.11
PyMuPDF>=1.23.0

# Machine learning & embeddings
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.1.0
scikit-learn>=1.3.0

# Vector search & data
faiss-cpu>=1.7.4
pandas>=2.1.0
numpy>=1.24.0

# API clients
requests>=2.31.0

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
jupyter>=1.0.0